refactor: move components to containers :+1:
https://redux.js.org/basics/usage-with-react#presentational-and-container-components
