+++ src/components/App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import './App.css';
-import { BooksForm } from './BooksForm';
-import BooksList from './BooksList';
+import { BooksForm } from '../containers/BooksForm';
+import BooksList from '../containers/BooksList';
 
 function App() {
   return (
     <div className="App">
-      <BooksList />
       <BooksForm />
+      <BooksList />
     </div>
   );
 }
